requirements.txt
fastapi
uvicorn
playwright
jinja2
python-multipart
pydantic
beautifulsoup4
aiofiles
tldextract
lxml
dnspython
aiohttp
pydantic-settings
olefile
